Bad request exception

+++ src/main/java/com/jlucas/dscommerce/services/ProductService.java
@@ -1,15 +1,20 @@
 package com.jlucas.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.jlucas.dscommerce.dto.ProductDTO;
 import com.jlucas.dscommerce.entities.Product;
+import com.jlucas.dscommerce.exceptions.DatabaseException;
 import com.jlucas.dscommerce.exceptions.ResourceNotFoundException;
 import com.jlucas.dscommerce.repositories.ProductRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 
@@ -46,15 +51,28 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto){
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado.");
+		}
+	
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id){
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
+		
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
